moved download package to api/v1/packages/download
