+++ fedn/network/storage/statestore/mongostatestore.py
@@ -98,21 +98,6 @@ class MongoStateStore:
         """
         return StringToReducerState(self.state.find_one()["current_state"])
 
-    def get_compute_package(self):
-        """Get the active compute package.
-
-        :return: The active compute package.
-        :rtype: ObjectID
-        """
-        try:
-            find = {"key": "active"}
-            projection = {"key": False, "_id": False}
-            ret = self.control.package.find_one(find, projection)
-            return ret
-        except Exception as e:
-            logger.error("ERROR: {}".format(e))
-            return None
-
     def get_storage_backend(self):
         """Get the storage backend.
